# perf: convert registry to JSON