Close #2517

## Benchmark

```console
$ hyperfine -N --warmup 3 '/Users/shunsukesuzuki/.local/share/aquaproj-aqua/internal/pkgs/github_release/github.com/aquaproj/aqua/v2.20.0/aqua_darwin_arm64.tar.gz/aqua exec -- goreleaser -v' '/Users/shunsukesuzuki/go/bin/aqua exec -- goreleaser -v'
Benchmark 1: /Users/shunsukesuzuki/.local/share/aquaproj-aqua/internal/pkgs/github_release/github.com/aquaproj/aqua/v2.20.0/aqua_darwin_arm64.tar.gz/aqua exec -- goreleaser -v
  Time (mean ± σ):      75.1 ms ±   0.9 ms    [User: 36.5 ms, System: 12.5 ms]
  Range (min … max):    73.6 ms …  78.3 ms    39 runs
 
Benchmark 2: /Users/shunsukesuzuki/go/bin/aqua exec -- goreleaser -v
  Time (mean ± σ):      50.4 ms ±   0.8 ms    [User: 30.1 ms, System: 12.4 ms]
  Range (min … max):    48.7 ms …  52.4 ms    58 runs
 
Summary
  /Users/shunsukesuzuki/go/bin/aqua exec -- goreleaser -v ran
    1.49 ± 0.03 times faster than /Users/shunsukesuzuki/.local/share/aquaproj-aqua/internal/pkgs/github_release/github.com/aquaproj/aqua/v2.20.0/aqua_darwin_arm64.tar.gz/aqua exec -- goreleaser -v
```